+++ app/javascript/components/Results/ResultGrid/ResultGrid.tsx
@@ -1,8 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { GridContainer, Grid } from '@trussworks/react-uswds';
 import parse from 'html-react-parser';
 import { clickTracking, truncateUrl } from '../../../utils';
 import { moduleCode } from '../../../utils/constants';
+import ResultGridWrapper from './ResultGridWrapper';
+import ResultTitle from './ResultTitle';
 
 type Result = {
   title: string,
@@ -54,34 +56,11 @@ export const ResultGrid = ({ result, affiliate, query, position, vertical }: Res
     return moduleCode[moduleKey as keyof typeof moduleCode];
   })();
 
-  const [isResultDivClickable, setIsResultDivClickable] = useState(false);
-  const [mobileResultDivStyle, setMobileResultDivStyle] = useState('');
-  
-  const isMobile = () => {
-    return window.innerWidth <= 480;
-  };
-  
-  const handleResultDivClick = (affiliate: string, module: string, query: string, position: number, url: string, vertical: string) => {
-    if (isResultDivClickable) {
-      clickTracking(affiliate, module, query, position, url, vertical);
-      if (mobileResultDivStyle !== '') 
-        setMobileResultDivStyle('');
-      else 
-        setMobileResultDivStyle('mobile-outline');
-      window.location.href = url;
-    }
-  };
-
-  useEffect(() => {
-    setIsResultDivClickable(isMobile());
-  }, []);
-
   return (
     <GridContainer className='result search-result-item'>
-      <Grid 
-        row gap="md" 
-        onClick={() => handleResultDivClick(affiliate, module, query, position, result.url, vertical)}
-        className={mobileResultDivStyle}>
+      <ResultGridWrapper
+        url={result.url}
+        clickTracking={()=> clickTracking(affiliate, module, query, position, result.url, vertical)}>
         {result.thumbnailUrl &&
         <Grid mobileLg={{ col: 4 }} className='result-thumbnail'>
           <img src={result.thumbnailUrl} className="result-image" alt={result.title}/>
@@ -93,12 +72,12 @@ export const ResultGrid = ({ result, affiliate, query, position, vertical }: Res
           {result.publishedAt   && <span className='published-date'>{result.publishedAt}</span>}
           <div className='result-title'>
             <h2 className='result-title-label'>
-              <a href={result.url}
-                className='result-title-link'
-                onClick={() => clickTracking(affiliate, module, query, position, result.url, vertical)}>
+              <ResultTitle 
+                url={result.url}
+                clickTracking={()=> clickTracking(affiliate, module, query, position, result.url, vertical)}>
                 {parse(result.title)} 
                 {getFileType(result.fileType)}
-              </a>
+              </ResultTitle>
             </h2>
           </div>
           <div className='result-desc'>
@@ -106,7 +85,7 @@ export const ResultGrid = ({ result, affiliate, query, position, vertical }: Res
             <div className='result-url-text'>{truncateUrl(result.url, URL_LENGTH)}</div>
           </div>
         </Grid>
-      </Grid>
+      </ResultGridWrapper>
       <Grid row className="row-mobile-divider"></Grid>
     </GridContainer>
   );
